pyproject.toml
[project]
name = "yt-to-knowledge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ollama>=0.6.1",
    "openai-whisper>=20250625",
    "yt-dlp>=2026.1.29",
]
